requirements.txt
torch>=2.2.1
transformers>=4.47.0
Pillow>=10.2.0
